✨ feat: 세션 데이터 테이블 구현
