--- src/components/molecules/chart/ChartCollection.tsx
@@ -1,36 +1,0 @@
-import { ChartWrapper } from 'types/ChartData';
-import ActiveTotalSessionChart from './ActiveTotalSessionChart';
-import BlockingWaitSessionChart from './BlockingWaitSessionChart';
-import SessionCountGroupChart from './SessionCountGroupChart';
-
-const ChartCollection = ({ chartData }: ChartWrapper) => {
-  return (
-    <div className="dashboard__charts">
-      <div className="dashboard__main-chart">
-        <ActiveTotalSessionChart
-          labels={chartData.labels}
-          totalSessions={chartData.totalSessions}
-          activeSessions={chartData.activeSessions}
-        />
-        <BlockingWaitSessionChart
-          labels={chartData.labels}
-          blockingSessions={chartData.blockingSessions}
-          waitSessions={chartData.waitSessions}
-        />
-      </div>
-
-      <div className="dashboard__sub-chart">
-        {['User', 'Type', 'Command', 'Machine'].map((group) => (
-          <SessionCountGroupChart
-            key={group}
-            labels={chartData.labels}
-            groupData={chartData[`sessionCountGroupBy${group}`]}
-            groupName={group}
-          />
-        ))}
-      </div>
-    </div>
-  );
-};
-
-export default ChartCollection;
