+++ src/services/deviceMonitoring.ts
@@ -3,7 +3,7 @@ import { apiInstance } from 'services';
 
 const api = apiInstance();
 
-const getMonitoringData = async (
+const getChartData = async (
   deviceAlias: string,
   interval: string,
   startTime: string,
@@ -29,6 +29,28 @@ const getMonitoringData = async (
     .catch(fail);
 };
 
+const getGridData = async (
+  deviceAlias: string,
+  collectedAt: string,
+  success: (response: AxiosResponse) => void,
+  fail: (error: AxiosError) => void
+): Promise<void> => {
+  const token = localStorage.getItem('accessToken');
+  await api({
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+    method: 'get',
+    url: '/monitoring/grid',
+    params: {
+      deviceAlias,
+      collectedAt,
+    },
+  })
+    .then(success)
+    .catch(fail);
+};
+
 const getSummaryData = async (
   success: (response: AxiosResponse) => void,
   fail: (error: AxiosError) => void
@@ -43,4 +65,4 @@ const getSummaryData = async (
     .catch(fail);
 };
 
-export { getMonitoringData, getSummaryData };
+export { getChartData, getGridData, getSummaryData };
